requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_s3_assets
aws_cdk.aws_iam